Data_Session2/Part-1/Deployement_Azure/Deployement_Code/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
azure-storage-blob
requests
pandas
